requirements.txt
einops==0.4.1
kornia==0.6.5
pytorch-lightning==1.5.9
matplotlib==3.5.2
opencv-python==4.6.0.66
lpips
imageio
imageio-ffmpeg
jupyter
scipy

# git+https://github.com/openai/CLIP.git
